requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
pytest==8.2.2
pytest-mock==3.14.0
python-keycloak
python-dotenv
Pillow
pytest-asyncio
email-validator
pydantic[email]